gfx/layers/apz/test/mochitest/test_group_touchevents-4.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Various touch tests that spawn in new windows (4)</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript" src="apz_test_native_event_utils.js"></script>
  <script type="application/javascript" src="apz_test_utils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript">

var touch_action_prefs = getPrefs("TOUCH_ACTION");

var subtests = [
  // clicking on element with :active::after CSS property
  {"file": "helper_bug1473108.html"},
  // Resetting isFirstPaint shouldn't clobber the visual viewport
  {"file": "helper_bug1509575.html", "prefs": getPrefs("TOUCH_EVENTS:PAN")},
  // Exercise one of the main-thread touch-action determination codepaths.
  {"file": "helper_bug1506497_touch_action_fixed_on_fixed.html", "prefs": touch_action_prefs},
  {"file": "helper_bug1637113_main_thread_hit_test.html"},
  {"file": "helper_bug1638458_contextmenu.html"},
  {"file": "helper_bug1638441_fixed_pos_hit_test.html"},
  {"file": "helper_bug1637135_narrow_viewport.html", "prefs": [["dom.meta-viewport.enabled", true]]},
  {"file": "helper_bug1714934_mouseevent_buttons.html"},
  {"file": "helper_bug1889017_fullscreen.html", "prefs": [
    ["dom.meta-viewport.enabled", true],
    // For the test to exercise the intended bug, this needs to be less than
    // the browser width divided by the device scale.
    ["browser.viewport.desktopWidth", 600],
    ["full-screen-api.allow-trusted-requests-only", false]
  ]},

  // Add new subtests to test_group_touch_events-N.html (for largest N),
  // not this file.
];

if (isApzEnabled()) {
  ok(window.TouchEvent, "Check if TouchEvent is supported (it should be, the test harness forces it on everywhere)");
  if (getPlatform() == "android") {
    // This has a lot of subtests, and Android emulators are slow.
    SimpleTest.requestLongerTimeout(2);
  }

  SimpleTest.waitForExplicitFinish();
  window.onload = function() {
    runSubtestsSeriallyInFreshWindows(subtests)
    .then(SimpleTest.finish, SimpleTest.finishWithFailure);
  };
}

  </script>
</head>
<body>
</body>
</html>
